+++ packages/space-app-components/src/recipes/browser.ts
@@ -1,0 +1,5 @@
+export * from "./index.js";
+
+import { defineSpaceChatElements } from "../chat/elements.js";
+
+defineSpaceChatElements();