+++ packages/space-app-components/src/recipes/index.ts
@@ -1,0 +1,15 @@
+export {
+  mountChatDrawerRecipe,
+  mountDefaultChatRecipe,
+  resolveSpaceChatRecipeElements,
+  type SpaceChatRecipeCopy,
+  type SpaceChatRecipeElements,
+  type SpaceChatRecipeHandle,
+  type SpaceChatRecipeMode,
+  type SpaceChatRecipeOptions,
+} from "./chat.js";
+export {
+  createSpaceComponentContext,
+  type SpaceComponentContext,
+  type SpaceComponentContextOptions,
+} from "../core/context.js";